# examples/vulnerable_demo_api/openapi.yaml
openapi: 3.1.0
info:
  title: Vulnerable Demo API
  version: 0.1.0
paths:
...
  /v1/projects/{projectId}:
    delete:
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
...